This is way better, Why would salavat copy the same snippet everywhere?
