+++ .gitignore
@@ -8,5 +8,3 @@
 /captures
 .externalNativeBuild
 /gradle
-test
-androidTest
