+++ .idea/modules.xml
@@ -2,8 +2,8 @@
 <project version="4">
   <component name="ProjectModuleManager">
     <modules>
-      <module fileurl="file://$PROJECT_DIR$/Correctly.iml" filepath="$PROJECT_DIR$/Correctly.iml" />
       <module fileurl="file://$PROJECT_DIR$/app/app.iml" filepath="$PROJECT_DIR$/app/app.iml" />
+      <module fileurl="file://$PROJECT_DIR$/correctly.iml" filepath="$PROJECT_DIR$/correctly.iml" />
     </modules>
   </component>
 </project>
